float

+++ lab2/example1.py
@@ -1,4 +1,4 @@
-x = int(input("Please input a number:"))
+x = float(input("Please input a number:"))
 
 if x >= 0:
   print("The absolute value of this number is" , x )
